+++ spec/support/article_representer.rb
@@ -1,6 +1,6 @@
 module ArticleRepresenter
-  include Roar::Representer::JSON
-  include Roar::Representer::Feature::Hypermedia
+  include Roar::JSON
+  include Roar::Hypermedia
 
   property :title
   property :id