+++ spec/support/order_representer.rb
@@ -1,8 +1,8 @@
 require 'support/article'
 
 module OrderRepresenter
-  include Roar::Representer::JSON
-  include Roar::Representer::Feature::Hypermedia
+  include Roar::JSON
+  include Roar::Hypermedia
 
   property :id
   property :client_id